Move p4runtime lib to stratum/lib
