+++ stratum/lib/p4runtime/entity_management.h
