+++ stratum/tools/benchmark/BUILD
@@ -16,63 +16,11 @@ package(
     default_visibility = STRATUM_INTERNAL,
 )
 
-stratum_cc_library(
-    name = "entity_management",
-    srcs = ["entity_management.cc"],
-    hdrs = ["entity_management.h"],
-    deps = [
-        "//stratum/glue/gtl:map_util",
-        "//stratum/glue/status",
-        "//stratum/lib:macros",
-        "//stratum/lib:utils",
-        "@com_github_p4lang_p4runtime//:p4info_cc_proto",
-        "@com_github_p4lang_p4runtime//:p4runtime_cc_proto",
-        "@com_google_absl//absl/container:flat_hash_map",
-        "@com_google_absl//absl/strings",
-    ],
-)
-
-stratum_cc_library(
-    name = "p4runtime_session",
-    srcs = ["p4runtime_session.cc"],
-    hdrs = ["p4runtime_session.h"],
-    deps = [
-        "//stratum/glue/status",
-        "//stratum/glue/status:statusor",
-        "//stratum/hal/lib/p4:utils",
-        "//stratum/lib:utils",
-        "//stratum/public/lib:error",
-        "@com_github_grpc_grpc//:grpc++",
-        "@com_github_grpc_grpc//:grpc++_public_hdrs",
-        "@com_github_p4lang_p4runtime//:p4runtime_cc_grpc",
-        "@com_github_p4lang_p4runtime//:p4runtime_cc_proto",
-        "@com_google_absl//absl/memory",
-        "@com_google_absl//absl/numeric:int128",
-        "@com_google_absl//absl/status:statusor",
-        "@com_google_absl//absl/time",
-    ],
-)
-
-stratum_cc_library(
-    name = "p4runtime_fixture",
-    testonly = 1,
-    # srcs = ["p4runtime_fixture.cc"],
-    hdrs = ["p4runtime_fixture.h"],
-    deps = [
-        ":p4runtime_session",
-        "//stratum/lib:utils",
-        "@com_github_google_glog//:glog",
-        "@com_google_googletest//:gtest",
-    ],
-)
-
 stratum_cc_binary(
     name = "stratum_benchmark",
     srcs = ["stratum_benchmark.cc"],
     arches = HOST_ARCHES,
     deps = [
-        ":entity_management",
-        ":p4runtime_session",
         "//stratum/glue/gtl:map_util",
         "//stratum/glue/status",
         "//stratum/glue/status:status_macros",
@@ -80,6 +28,8 @@ stratum_cc_binary(
         "//stratum/hal/lib/p4:utils",
         "//stratum/lib:macros",
         "//stratum/lib:utils",
+        "//stratum/lib/p4runtime:entity_management",
+        "//stratum/lib/p4runtime:p4runtime_session",
         "@com_github_gflags_gflags//:gflags",
         "@com_github_grpc_grpc//:grpc++",
         "@com_github_p4lang_p4runtime//:p4info_cc_proto",
@@ -96,14 +46,14 @@ cc_test(
     linkstatic = True,
     tags = ["manual"],
     deps = [
-        ":p4runtime_fixture",
-        ":p4runtime_session",
         "//stratum/glue/status",
         "//stratum/glue/status:status_macros",
         "//stratum/glue/status:status_test_util",
         "//stratum/lib:macros",
         "//stratum/lib:test_main",
         "//stratum/lib:utils",
+        "//stratum/lib/p4runtime:p4runtime_fixture",
+        "//stratum/lib/p4runtime:p4runtime_session",
         "//stratum/lib/test_utils:matchers",
         "//stratum/public/lib:error",
         "@com_github_gflags_gflags//:gflags",
